refactor(contracts): share secret masking helpers
