+++ apps/server/src/utils/secretMasking.ts
@@ -1,5 +1,7 @@
-const FULL_MASK = '****';
-const VISIBLE_CHARS = 3;
+import {
+  maskSecret as sharedMaskSecret,
+  maskSecretString as sharedMaskSecretString,
+} from '@maintainerr/contracts';
 
 const SENSITIVE_PARAM_KEYS = [
   'api_key',
@@ -249,24 +251,9 @@ const sanitizePathSegmentValues = (value: string, segment: string): string => {
   return sanitized;
 };
 
-export const maskSecret = (value: string | null | undefined): string | null => {
-  if (value == null) {
-    return null;
-  }
+export const maskSecret = sharedMaskSecret;
 
-  if (value === '') {
-    return '';
-  }
-
-  if (value.length <= VISIBLE_CHARS * 2) {
-    return FULL_MASK;
-  }
-
-  return `${value.slice(0, VISIBLE_CHARS)}...${value.slice(-VISIBLE_CHARS)}`;
-};
-
-export const maskSecretString = (value: string): string =>
-  maskSecret(value) ?? FULL_MASK;
+export const maskSecretString = sharedMaskSecretString;
 
 export const sanitizeSecretString = (value: string): string => {
   let sanitized = sanitizeAuthorizationValues(value);
