+++ packages/contracts/src/app/index.ts
@@ -1,2 +1,3 @@
 export * from './basicResponse.dto'
+export * from './maskSecret'
 export * from './versionResponse.dto'
